+++ src/test-utils.tsx
@@ -22,6 +22,7 @@ export class MockedProvider extends React.Component<any, any> {
 
   constructor(props, context) {
     super(props, context);
+    if (this.props.client) return;
 
     const networkInterface = mockNetworkInterface.apply(null, this.props.mocks);
     this.client = new ApolloClient({ networkInterface });
@@ -29,7 +30,7 @@ export class MockedProvider extends React.Component<any, any> {
 
   render() {
     return (
-      <ApolloProvider client={this.client}>
+      <ApolloProvider client={this.client || this.props.client} store={this.props.store || null}>
         {this.props.children}
       </ApolloProvider>
     );
@@ -106,11 +107,11 @@ export class MockNetworkInterface implements NetworkInterface {
       if (!mockedResponse.result && !mockedResponse.error) {
         throw new Error('Mocked response should contain either result or error.');
       }
-      this.addMockedReponse(mockedResponse);
+      this.addMockedResponse(mockedResponse);
     });
   }
 
-  public addMockedReponse(mockedResponse: MockedResponse) {
+  public addMockedResponse(mockedResponse: MockedResponse) {
     const key = requestToKey(mockedResponse.request);
     let mockedResponses = this.mockedResponsesByKey[key];
     if (!mockedResponses) {